Main  Menu Java File
This is an example of the main menu file that will call on the functions of the other programs. 

Put the name of the games where Lines 9 - 13 are. 

Note the format of Lines  57, 67, 77, 87, and 97. Substitute "<NAMEOFGAMEFILE>" with the main class of each game and take out the "//". It will call the main method of the game file